Refactor!: redesign the application as a multi-page app
