+++ app.py
@@ -1,4 +1,5 @@
 # dash
+import dash
 import dash_mantine_components as dmc
 from dash import Dash, dcc, html
 
@@ -6,7 +7,7 @@ from dash import Dash, dcc, html
 from dotenv import load_dotenv; load_dotenv()
 
 # local packages
-from src import callbacks, ui
+from src import callbacks
 
 # app definition
 dash_app = Dash(
@@ -20,26 +21,24 @@ dash_app = Dash(
     suppress_callback_exceptions=True,
     title='SDG App',
     update_title='Loading...',
+    use_pages=True,
 )
 app = dash_app.server
 
 # define layout
 dash_app.layout = dmc.NotificationsProvider(
     children=[
-        dcc.Store(id='user-config', storage_type='memory'),
-        dcc.Store(id='session-config', storage_type='memory'),
+        dcc.Location(id='location'),
+        dcc.Store(id='user-config', storage_type='session'),
+        dcc.Store(id='session-config', storage_type='session'),
         dcc.Interval(
             id='interval-component',
             interval=1_000 * 10,  # in milliseconds
             n_intervals=0
         ),
-        dmc.Grid(
-            id='content',
-            children=ui.get_start_layout(),
-            justify='center',
-            gutter='sm',
-        ),
+        dash.page_container,
         html.Div(id='notifications-container'),
+        html.Div(id='login-redirect'),
     ],
     autoClose=False,
     position='top-center',
