+++ src/callbacks/__init__.py
@@ -1,0 +1,1 @@
+from . import header, labelling, login, profile, quit, setup
