+++ src/ui/__init__.py
@@ -1,2 +1,1 @@
-from . import affixes, header, footer, layouts, buttons, extras, styles, notifications, accordions, tables
-from .layouts import get_start_layout, get_main_layout, get_finish_layout
+from . import alerts, affixes, header, footer, layouts, buttons, extras, styles, notifications, accordions, tables, inputs
